+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/acceptance/privacy/EnclaveErrorAcceptanceTest.java
@@ -49,7 +49,6 @@ import org.web3j.protocol.besu.response.privacy.PrivateTransactionReceipt;
 import org.web3j.utils.Restriction;
 
 @RunWith(Parameterized.class)
-@Ignore("Ignored since Tessera/Docker container startup causing errors")
 public class EnclaveErrorAcceptanceTest extends PrivacyAcceptanceTestBase {
 
   private final PrivacyNode alice;