+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/acceptance/privacy/PrivGetLogsAcceptanceTest.java
@@ -30,14 +30,12 @@ import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.web3j.protocol.besu.response.privacy.PrivateTransactionReceipt;
 import org.web3j.protocol.core.methods.response.EthLog.LogResult;
 import org.web3j.utils.Restriction;
 
 @SuppressWarnings("rawtypes")
-@Ignore("Ignored since Tessera/Docker container startup causing errors")
 public class PrivGetLogsAcceptanceTest extends ParameterizedEnclaveTestBase {
 
   /*