+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 22.7.2
 
 ### Additions and Improvements
+- Upgrade besu-native to 0.6.0 and use Blake2bf native implementation if available by default [#4264](https://github.com/hyperledger/besu/pull/4264)
+- Better management of jemalloc presence/absence in startup script [#4237](https://github.com/hyperledger/besu/pull/4237)
+- Filter out disconnected peers when fetching available peers [#4269](https://github.com/hyperledger/besu/pull/4269)
 
 ### Bug Fixes
 
@@ -33,7 +36,7 @@
 ### Additions and Improvements
 - Deprecation warning for Ropsten, Rinkeby, Kiln [#4173](https://github.com/hyperledger/besu/pull/4173)
 
-### Bug Fixes 
+### Bug Fixes
 
 - Fixes previous known issue [#3890](https://github.com/hyperledger/besu/issues/3890)from RC3 requiring a restart post-merge to continue correct transaction handling.
 - Stop producing stack traces when a get headers response only contains the range start header [#4189](https://github.com/hyperledger/besu/pull/4189)
@@ -52,8 +55,8 @@
 ### Additions and Improvements
 - Engine API: Change expiration time for JWT tokens to 60s [#4168](https://github.com/hyperledger/besu/pull/4168)
 - Sepolia mergeNetSplit block [#4158](https://github.com/hyperledger/besu/pull/4158)
-- Goerli TTD [#4160](https://github.com/hyperledger/besu/pull/4160) 
-- Several logging improvements 
+- Goerli TTD [#4160](https://github.com/hyperledger/besu/pull/4160)
+- Several logging improvements
 
 ### Bug Fixes
 - Allow to set any value for baseFeePerGas in the genesis file [#4177](https://github.com/hyperledger/besu/pull/4177)